+++ src/include/sll/types.h
@@ -536,9 +536,9 @@ typedef struct __SLL_RUNTIME_DATA{
 
 
 typedef struct __SLL_RUNTIME_OBJECT_STACK_DATA{
-	uint32_t off;
 	uint64_t* m;
 	uint32_t ml;
+	sll_ref_count_t* s;
 } sll_runtime_object_stack_data_t;