+++ src/sll/free.c
@@ -90,10 +90,11 @@ __SLL_FUNC void sll_free_internal_function_table(sll_internal_function_table_t* 
 
 
 __SLL_FUNC void sll_free_runtime_object_stack_data(sll_runtime_object_stack_data_t* rst){
-	rst->off=0;
 	free(rst->m);
 	rst->m=NULL;
 	rst->ml=0;
+	free(rst->s);
+	rst->s=NULL;
 }